requirements.txt
aiogram>=2.11.2,<3.0
aiohttp==3.7.4
python-dotenv
gino
asyncio
sqlalchemy
aioredis
datetime
typing
logging